+++ mayan/apps/document_indexing/apps.py
@@ -32,12 +32,12 @@ from .html_widgets import (
 from .links import (
     link_document_index_instance_list, link_document_type_index_templates,
     link_index_instance_menu, link_index_instance_rebuild,
-    link_index_template_setup, link_index_template_create,
-    link_index_template_document_types, link_index_template_delete,
-    link_index_template_edit, link_index_template_list,
-    link_index_template_node_tree_view, link_index_instances_rebuild,
-    link_index_template_node_create, link_index_template_node_delete,
-    link_index_template_node_edit
+    link_index_instances_reset, link_index_template_setup,
+    link_index_template_create, link_index_template_document_types,
+    link_index_template_delete, link_index_template_edit,
+    link_index_template_list, link_index_template_node_tree_view,
+    link_index_instances_rebuild, link_index_template_node_create,
+    link_index_template_node_delete, link_index_template_node_edit
 )
 from .permissions import (
     permission_document_indexing_create, permission_document_indexing_delete,
@@ -220,7 +220,9 @@ class DocumentIndexingApp(MayanAppConfig):
             )
         )
         menu_setup.bind_links(links=(link_index_template_setup,))
-        menu_tools.bind_links(links=(link_index_instances_rebuild,))
+        menu_tools.bind_links(
+            links=(link_index_instances_rebuild, link_index_instances_reset)
+        )
 
         post_delete.connect(
             dispatch_uid='document_indexing_handler_delete_empty',