+++ mayan/apps/document_indexing/links.py
@@ -37,6 +37,12 @@ link_index_instance_menu = Link(
     ), icon_class_path='mayan.apps.document_indexing.icons.icon_index',
     text=_('Indexes'), view='indexing:index_list'
 )
+link_index_instance_rebuild = Link(
+    args='resolved_object.pk',
+    icon_class_path='mayan.apps.document_indexing.icons.icon_index_instances_rebuild',
+    permissions=(permission_document_indexing_rebuild,),
+    text=_('Rebuild index'), view='indexing:index_setup_rebuild'
+)
 link_index_instances_rebuild = Link(
     condition=get_cascade_condition(
         app_label='document_indexing', model_name='Index',
@@ -48,11 +54,16 @@ link_index_instances_rebuild = Link(
     ),
     text=_('Rebuild indexes'), view='indexing:rebuild_index_instances'
 )
-link_index_instance_rebuild = Link(
-    args='resolved_object.pk',
-    icon_class_path='mayan.apps.document_indexing.icons.icon_index_instances_rebuild',
-    permissions=(permission_document_indexing_rebuild,),
-    text=_('Rebuild index'), view='indexing:index_setup_rebuild'
+link_index_instances_reset = Link(
+    condition=get_cascade_condition(
+        app_label='document_indexing', model_name='Index',
+        object_permission=permission_document_indexing_rebuild,
+    ),
+    icon_class_path='mayan.apps.document_indexing.icons.icon_index_instances_reset',
+    description=_(
+        'Deletes and creates from scratch all the document indexes.'
+    ),
+    text=_('Reset indexes'), view='indexing:index_instances_reset'
 )
 
 link_index_template_setup = Link(