+++ mayan/apps/document_indexing/views.py
@@ -474,3 +474,39 @@ class IndexesRebuildView(FormView):
 
     def get_post_action_redirect(self):
         return reverse(viewname='common:tools_list')
+
+
+class IndexesResetView(FormView):
+    extra_context = {
+        'title': _('Reset indexes'),
+    }
+    form_class = IndexTemplateFilteredForm
+
+    def form_valid(self, form):
+        count = 0
+        for index in form.cleaned_data['index_templates']:
+            index.instance_root.delete()
+            count += 1
+
+        messages.success(
+            message=ungettext(
+                singular='%(count)d index reset.',
+                plural='%(count)d indexes reset.',
+                number=count
+            ) % {
+                'count': count,
+            }, request=self.request
+        )
+
+        return super(IndexesResetView, self).form_valid(form=form)
+
+    def get_form_extra_kwargs(self):
+        return {
+            'help_text': _(
+                'Index templates for which their instances will be deleted.'
+            ),
+            'user': self.request.user
+        }
+
+    def get_post_action_redirect(self):
+        return reverse(viewname='common:tools_list')